nginx: make it work on read-only fs
By default, nginx writes to various files, mainly for logging and
caching. In order for it to work with read-only file systems, this patch
provides an alternative nginx configuration file redirecting all these
write paths to /tmp (which is backed by ramfs and thus is read-write).

On the way, nginx is also updated to 1.19.7 (current latest).